rename: 안 쓰는 코드, 안 쓰는 파일 삭제에 따른 변경
